Incorrect styles when items selected or highlighted #435
-need to adjust install app dialog grid css
