+++ src/main/resources/assets/js/app/browse/ApplicationTreeGrid.ts
@@ -7,12 +7,10 @@ import {ApplicationKey} from 'lib-admin-ui/application/ApplicationKey';
 import {i18n} from 'lib-admin-ui/util/Messages';
 import {TreeGrid} from 'lib-admin-ui/ui/treegrid/TreeGrid';
 import {TreeGridBuilder} from 'lib-admin-ui/ui/treegrid/TreeGridBuilder';
-import {ResponsiveRanges} from 'lib-admin-ui/ui/responsive/ResponsiveRanges';
 import {TreeGridContextMenu} from 'lib-admin-ui/ui/treegrid/TreeGridContextMenu';
 import {GetApplicationRequest} from 'lib-admin-ui/application/GetApplicationRequest';
 import {DefaultErrorHandler} from 'lib-admin-ui/DefaultErrorHandler';
 import {ListApplicationsRequest} from 'lib-admin-ui/application/ListApplicationsRequest';
-import {Body} from 'lib-admin-ui/dom/Body';
 import {UploadItem} from 'lib-admin-ui/ui/uploader/UploadItem';
 
 declare const CONFIG;
@@ -39,44 +37,9 @@ export class ApplicationTreeGrid extends TreeGrid<Application> {
             style: {cssClass: 'state', minWidth: 80, maxWidth: 100}
         }]).prependClasses('application-grid');
 
-        const columns = builder.getColumns().slice(0);
-        const [
-            nameColumn,
-            versionColumn,
-            stateColumn,
-        ] = columns;
-
-        const updateColumns = () => {
-            let width = this.getEl().getWidth();
-            let checkSelIsMoved = ResponsiveRanges._540_720.isFitOrSmaller(Body.get().getEl().getWidth());
-
-            const curClass = nameColumn.getCssClass();
-
-            if (checkSelIsMoved) {
-                nameColumn.setCssClass(curClass || 'shifted');
-            } else if (curClass && curClass.indexOf('shifted') >= 0) {
-                nameColumn.setCssClass(curClass.replace('shifted', ''));
-            }
-
-            if (ResponsiveRanges._240_360.isFitOrSmaller(width)) {
-                nameColumn.setBoundaryWidth(150, 250);
-                versionColumn.setBoundaryWidth(50, 70);
-                stateColumn.setBoundaryWidth(50, 50);
-            } else if (ResponsiveRanges._360_540.isFitOrSmaller(width)) {
-                nameColumn.setBoundaryWidth(200, 350);
-                versionColumn.setBoundaryWidth(50, 70);
-                stateColumn.setBoundaryWidth(50, 70);
-            } else {
-                nameColumn.setBoundaryWidth(200, 9999);
-                versionColumn.setBoundaryWidth(50, 130);
-                stateColumn.setBoundaryWidth(80, 100);
-            }
-            this.setColumns(columns.slice(0), checkSelIsMoved);
-        };
-
         const readonlyMode: boolean = CONFIG.readonlyMode === 'true';
+
         builder.setCheckableRows(!readonlyMode);
-        builder.setColumnUpdater(updateColumns);
 
         super(builder);
