+++ src/main/resources/assets/styles/installation/install-application-dialog.less
@@ -82,103 +82,6 @@
 
     .modal-dialog-body {
       padding-top: 0;
-
-      .dialog-content {
-        .market-apps {
-          .market-app-tree-grid {
-            .grid {
-
-              .slick-cell {
-                padding-top: 15px;
-                height: 52px;
-                cursor: default;
-
-                * {
-                  cursor: default;
-                }
-
-                .viewer {
-                  .names-and-icon-view {
-                    cursor: default;
-                  }
-                }
-
-                &.app-name-and-icon {
-                  margin-left: 10px;
-
-                  .@{_COMMON_PREFIX}main-name {
-                    font-size: 16px;
-                    padding-bottom: 2px;
-                    color: @admin-blue;
-
-                    a {
-                      cursor: pointer !important;
-                      text-decoration: none;
-
-                      &:hover, &:focus {
-                        text-decoration: underline;
-                      }
-                    }
-                  }
-                }
-
-                .toggle.icon {
-                  display: none;
-                }
-              }
-            }
-
-            .version {
-              text-align: center;
-              color: @admin-black;
-              font-size: 12px;
-            }
-
-            .app-status {
-              text-align: center;
-
-              a {
-                text-decoration: none;
-              }
-
-              .install, .update {
-                font-size: 16px;
-                cursor: pointer !important;
-
-                span {
-                  cursor: pointer !important;
-                }
-              }
-
-              .install {
-                &:hover, &:focus {
-                  text-decoration: underline;
-                }
-
-                color: @admin-green;
-              }
-
-              .update {
-                &:hover, &:focus {
-                  text-decoration: underline;
-                }
-
-                color: @input-dark-red;
-              }
-
-              .installing .progress-bar {
-                display: block;
-              }
-
-              .installed {
-                opacity: 0.6;
-                font-size: 14px;
-                color: @admin-font-gray2;
-              }
-            }
-          }
-        }
-      }
     }
   }
