+++ src/main/resources/assets/styles/main.less
@@ -6,6 +6,7 @@
 
 @import "installation/application-uploader-el";
 @import "installation/install-application-dialog";
+@import "installation/market-app-tree-grid";
 @import "browse/application-tree-grid";
 @import "new/new-schema-dialog";
 @import "view/application-browse-panel";
